Resolving Build errors
