+++ apps/e-shop-web/app/global-error.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+export const dynamic = 'force-dynamic';
+
+export default function GlobalError({
+  error,
+  reset,
+}: {
+  error: Error & { digest?: string };
+  reset: () => void;
+}) {
+  console.error(error);
+
+  return (
+    <html>
+      <body>
+        <div className="flex flex-col items-center justify-center min-h-screen bg-gray-100">
+          <h2 className="text-2xl font-bold text-gray-800 mb-4">
+            Something went wrong!
+          </h2>
+          <button
+            onClick={() => reset()}
+            className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
+          >
+            Try again
+          </button>
+        </div>
+      </body>
+    </html>
+  );
+}
