+++ apps/e-shop-web/app/not-found.tsx
@@ -1,0 +1,22 @@
+'use client';
+
+export const dynamic = 'force-dynamic';
+
+export default function NotFound() {
+  return (
+    <div className="flex flex-col items-center justify-center min-h-screen bg-gray-100">
+      <h1 className="text-4xl font-bold text-gray-800 mb-4">
+        404 - Page Not Found
+      </h1>
+      <p className="text-gray-600 mb-8">
+        The page you're looking for doesn't exist.
+      </p>
+      <a
+        href="/"
+        className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
+      >
+        Go Home
+      </a>
+    </div>
+  );
+}
