+++ apps/e-shop-web/eslint.config.mjs
@@ -7,6 +7,6 @@ export default [
   ...baseConfig,
   ...nx.configs['flat/react-typescript'],
   {
-    ignores: ['.next/**/*', '**/out-tsc'],
+    ignores: ['.next/**/*', '**/out-tsc', 'next-env.d.ts'],
   },
 ];
