+++ apps/e-shop-web/next.config.js
@@ -9,6 +9,10 @@ const nextConfig = {
   // Use this to set Nx-specific options
   // See: https://nx.dev/recipes/next/next-config-setup
   nx: {},
+  output: 'standalone',
+  experimental: {
+    serverComponentsExternalPackages: ['@tanstack/react-query'],
+  },
 };
 
 const plugins = [
